tests: Factor out test fixture as a separate module
OK so...buckle up.  I've been working on this ostree stuff for
like 10 years now it turns out that I still stuck at writing
the test suites.

When I generated the hardcoded `exampleos-v0.tar.zstd`, I apparently
included SELinux labels (I don't think intentionally).
But I didn't when generating `exampleos-v1.tar`!

I was trying to test an upgrade path around xattrs and hit a bug there
code, and it took me a while to understand it was because the second
update didn't have them!

The time it took to debug this was greatly extended by me reading
`man tar`, and thinking that `tar -tf -v --xattrs` would print any SELinux
label.  But no - SELinux is *special cased* in tar - even though
it's an extended attribute, in order to see any embedded labels
in the tarball one needs to use `tar -tf -v --selinux`!
Somehow I had apparently avoided learning this until now.  (ostree's
model I think is better than tar's, but here we're intersecting them)

Anyways.  None of this is fixed right now.  Yes, I can see the
disappointment.

But - in preparation for making the test suite more robust and
fixing things up, factor out the fixture as a separate module.
Make things like the testref a method instead of using a constant.

Port the diff test to use this too.
